+++ test/test-api.js
@@ -37,6 +37,27 @@ app.get("/redirect-target", (c) => {
   return c.json({ msg: "redirect-target" });
 });
 
+app.get("/users/:id", (c) => {
+  return c.json({ id: c.req.param("id") });
+});
+
+app.get("/echo-headers", (c) => {
+  const headers = {};
+  const excludeHeaders = [
+    "accept",
+    "accept-encoding",
+    "accept-language",
+    "host",
+    "user-agent",
+  ];
+  for (const [key, value] of c.req.raw.headers.entries()) {
+    if (!excludeHeaders.includes(key)) {
+      headers[key] = value;
+    }
+  }
+  return c.json(headers);
+});
+
 app.get("/auth-required", (c) => {
   if (c.req.header("Authorization") == "123") {
     return c.json({ id: "sample-post" });